Move RouterType to extractors/routers.rs
